# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API SAOARS/UB",
        "description": "Documentation de l'API API SAOARS/UB",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/abilities/{ability}": {
            "get": {
                "tags": [
                    "Ability"
                ],
                "summary": "Get a ability by slug",
                "operationId": "edeacf1594406853fdb4ecc799fe6800",
                "parameters": [
                    {
                        "name": "ability",
                        "in": "path",
                        "description": "Ability slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get a ability by slug",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Ability Name"
                                        },
                                        "slug": {
                                            "type": "string",
                                            "example": "ability-name"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Ability description"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ability not found"
                    }
                }
            }
        },
...
        "/api/weapons/{weapon}": {
            "get": {
                "tags": [
                    "Weapon"
                ],
                "summary": "Get a weapon by slug",
                "operationId": "d847555e1b2011721df0d04a8fb3794b",
                "parameters": [
                    {
                        "name": "weapon",
                        "in": "path",
                        "description": "Weapon slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get a weapon by slug",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Weapon Name"
                                        },
                                        "slug": {
                                            "type": "string",
                                            "example": "weapon-name"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Weapon description"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Weapon not found"
                    }
                }
            }
        }
    },
...
}